fix trialExpiration

+++ site/cmd/trialExpiration.php
@@ -1,5 +1,7 @@
 <?php
 
 $userIds = db()->selectCol('SELECT userId FROM invoice WHERE dueDate < ?', Date::db());
-db()->query("DELETE FROM sessions WHERE userId IN (?a)", $userIds);
-print "logged off ".count($userIds)." users\n";
+if ($userIds) {
+  db()->query("DELETE FROM sessions WHERE userId IN (?a)", $userIds);
+  print "logged off ".count($userIds)." users\n";
+}
